Feature: qutrit_ansatz() Set DTYPE back to torch.float32

+++ QudiTop/global_var.py
@@ -3,7 +3,8 @@ Refer: https://gitee.com/forcekeng/quditop
 """
 import torch
 
-DTYPE = torch.float64
+DTYPE = torch.float32
+# DTYPE = torch.float64
 CDTYPE = torch.complex128
 DEFAULT_VALUE = torch.tensor(0, dtype=DTYPE)
 DEFAULT_PARAM_NAME = '_param_'
